12 left!

+++ src/bun.js/node/node_fs.zig
@@ -32,8 +32,8 @@ const ArgumentsSlice = JSC.Node.ArgumentsSlice;
 const TimeLike = JSC.Node.TimeLike;
 const Mode = JSC.Node.Mode;
 
-const uid_t = std.os.uid_t;
-const gid_t = std.os.gid_t;
+const uid_t = if (Environment.isPosix) std.os.uid_t else i32;
+const gid_t = if (Environment.isPosix) std.os.gid_t else i32;
 /// u63 to allow one null bit
 const ReadPosition = i64;
 
@@ -1216,7 +1216,9 @@ pub const Arguments = struct {
                 // be absolute. When using 'junction', the target argument
                 // will automatically be normalized to absolute path.
                 if (next_val.isString()) {
-                    comptime if (Environment.isWindows) @compileError("Add support for type argument on Windows");
+                    if (comptime Environment.isWindows) {
+                        bun.todo(@src(), {});
+                    }
                     arguments.eat();
                 }
             }
@@ -2294,7 +2296,7 @@ pub const Arguments = struct {
                 .flag = flag,
                 .mode = mode,
                 .data = data,
-                .dirfd = @as(FileDescriptor, @intCast(std.fs.cwd().fd)),
+                .dirfd = bun.toFD(std.fs.cwd().fd),
             };
         }
     };
@@ -3063,6 +3065,10 @@ pub const NodeFS = struct {
     pub const ReturnType = Return;
 
     pub fn access(this: *NodeFS, args: Arguments.Access, comptime _: Flavor) Maybe(Return.Access) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Access).todo;
+        }
+
         var path = args.path.sliceZ(&this.sync_error_buf);
         const rc = Syscall.system.access(path, @intFromEnum(args.mode));
         return Maybe(Return.Access).errnoSysP(rc, .access, path) orelse Maybe(Return.Access).success;
@@ -3396,6 +3402,10 @@ pub const NodeFS = struct {
     }
 
     pub fn chown(this: *NodeFS, args: Arguments.Chown, comptime flavor: Flavor) Maybe(Return.Chown) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fchmod).todo;
+        }
+
         const path = args.path.sliceZ(&this.sync_error_buf);
 
         switch (comptime flavor) {
@@ -3408,6 +3418,10 @@ pub const NodeFS = struct {
 
     /// This should almost never be async
     pub fn chmod(this: *NodeFS, args: Arguments.Chmod, comptime flavor: Flavor) Maybe(Return.Chmod) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fchmod).todo;
+        }
+
         const path = args.path.sliceZ(&this.sync_error_buf);
 
         switch (comptime flavor) {
@@ -3423,6 +3437,10 @@ pub const NodeFS = struct {
 
     /// This should almost never be async
     pub fn fchmod(_: *NodeFS, args: Arguments.FChmod, comptime flavor: Flavor) Maybe(Return.Fchmod) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fchmod).todo;
+        }
+
         switch (comptime flavor) {
             .sync => {
                 return Syscall.fchmod(args.fd, args.mode);
@@ -3433,6 +3451,10 @@ pub const NodeFS = struct {
         return Maybe(Return.Fchmod).todo;
     }
     pub fn fchown(_: *NodeFS, args: Arguments.Fchown, comptime flavor: Flavor) Maybe(Return.Fchown) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fchown).todo;
+        }
+
         switch (comptime flavor) {
             .sync => {
                 return Maybe(Return.Fchown).errnoSys(C.fchown(args.fd, args.uid, args.gid), .fchown) orelse
@@ -3444,6 +3466,9 @@ pub const NodeFS = struct {
         return Maybe(Return.Fchown).todo;
     }
     pub fn fdatasync(_: *NodeFS, args: Arguments.FdataSync, comptime flavor: Flavor) Maybe(Return.Fdatasync) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fdatasync).todo;
+        }
         switch (comptime flavor) {
             .sync => return Maybe(Return.Fdatasync).errnoSys(system.fdatasync(args.fd), .fdatasync) orelse
                 Maybe(Return.Fdatasync).success,
@@ -3453,12 +3478,18 @@ pub const NodeFS = struct {
         return Maybe(Return.Fdatasync).todo;
     }
     pub fn fstat(_: *NodeFS, args: Arguments.Fstat, comptime flavor: Flavor) Maybe(Return.Fstat) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fstat).todo;
+        }
+
         switch (comptime flavor) {
             .sync => {
-                return switch (Syscall.fstat(args.fd)) {
-                    .result => |result| Maybe(Return.Fstat){ .result = Stats.init(result, false) },
-                    .err => |err| Maybe(Return.Fstat){ .err = err },
-                };
+                if (comptime Environment.isPosix) {
+                    return switch (Syscall.fstat(args.fd)) {
+                        .result => |result| Maybe(Return.Fstat){ .result = Stats.init(result, false) },
+                        .err => |err| Maybe(Return.Fstat){ .err = err },
+                    };
+                }
             },
             else => {},
         }
@@ -3467,6 +3498,10 @@ pub const NodeFS = struct {
     }
 
     pub fn fsync(_: *NodeFS, args: Arguments.Fsync, comptime flavor: Flavor) Maybe(Return.Fsync) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Fsync).todo;
+        }
+
         switch (comptime flavor) {
             .sync => return Maybe(Return.Fsync).errnoSys(system.fsync(args.fd), .fsync) orelse
                 Maybe(Return.Fsync).success,
@@ -3477,8 +3512,7 @@ pub const NodeFS = struct {
     }
 
     pub fn ftruncateSync(args: Arguments.FTruncate) Maybe(Return.Ftruncate) {
-        return Maybe(Return.Ftruncate).errnoSys(system.ftruncate(args.fd, args.len orelse 0), .ftruncate) orelse
-            Maybe(Return.Ftruncate).success;
+        return Syscall.ftruncate(args.fd, args.len orelse 0);
     }
 
     pub fn ftruncate(_: *NodeFS, args: Arguments.FTruncate, comptime flavor: Flavor) Maybe(Return.Ftruncate) {
@@ -3490,6 +3524,10 @@ pub const NodeFS = struct {
         return Maybe(Return.Ftruncate).todo;
     }
     pub fn futimes(_: *NodeFS, args: Arguments.Futimes, comptime flavor: Flavor) Maybe(Return.Futimes) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Futimes).todo;
+        }
+
         var times = [2]std.os.timespec{
             .{
                 .tv_sec = args.mtime,
@@ -3513,6 +3551,10 @@ pub const NodeFS = struct {
     }
 
     pub fn lchmod(this: *NodeFS, args: Arguments.LCHmod, comptime flavor: Flavor) Maybe(Return.Lchmod) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Lchmod).todo;
+        }
+
         const path = args.path.sliceZ(&this.sync_error_buf);
 
         switch (comptime flavor) {
@@ -3527,6 +3569,10 @@ pub const NodeFS = struct {
     }
 
     pub fn lchown(this: *NodeFS, args: Arguments.LChown, comptime flavor: Flavor) Maybe(Return.Lchown) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Lchown).todo;
+        }
+
         const path = args.path.sliceZ(&this.sync_error_buf);
 
         switch (comptime flavor) {
@@ -3555,6 +3601,10 @@ pub const NodeFS = struct {
         return Maybe(Return.Link).todo;
     }
     pub fn lstat(this: *NodeFS, args: Arguments.Lstat, comptime flavor: Flavor) Maybe(Return.Lstat) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Lstat).todo;
+        }
+
         _ = flavor;
         return switch (Syscall.lstat(
             args.path.sliceZ(
@@ -3594,7 +3644,7 @@ pub const NodeFS = struct {
     // TODO: verify this works correctly with unicode codepoints
     pub fn mkdirRecursive(this: *NodeFS, args: Arguments.Mkdir, comptime flavor: Flavor) Maybe(Return.Mkdir) {
         const Option = Maybe(Return.Mkdir);
-        if (comptime Environment.isWindows) @compileError("This needs to be implemented on Windows.");
+        if (comptime Environment.isWindows) return Option.todo;
 
         switch (comptime flavor) {
             // The sync version does no allocation except when returning the path
@@ -3827,6 +3877,7 @@ pub const NodeFS = struct {
     }
 
     pub fn read(this: *NodeFS, args: Arguments.Read, comptime flavor: Flavor) Maybe(Return.Read) {
+        if (comptime Environment.isWindows) return Maybe(Return.Read).todo;
         return if (args.position != null)
             this._pread(
                 args,
@@ -3840,14 +3891,17 @@ pub const NodeFS = struct {
     }
 
     pub fn readv(this: *NodeFS, args: Arguments.Readv, comptime flavor: Flavor) Maybe(Return.Read) {
+        if (comptime Environment.isWindows) return Maybe(Return.Read).todo;
         return if (args.position != null) _preadv(this, args, flavor) else _readv(this, args, flavor);
     }
 
     pub fn writev(this: *NodeFS, args: Arguments.Writev, comptime flavor: Flavor) Maybe(Return.Write) {
+        if (comptime Environment.isWindows) return Maybe(Return.Write).todo;
         return if (args.position != null) _pwritev(this, args, flavor) else _writev(this, args, flavor);
     }
 
     pub fn write(this: *NodeFS, args: Arguments.Write, comptime flavor: Flavor) Maybe(Return.Write) {
+        if (comptime Environment.isWindows) return Maybe(Return.Write).todo;
         return if (args.position != null) _pwrite(this, args, flavor) else _write(this, args, flavor);
     }
     fn _write(_: *NodeFS, args: Arguments.Write, comptime flavor: Flavor) Maybe(Return.Write) {
@@ -4028,7 +4082,7 @@ pub const NodeFS = struct {
         }
 
         var entries = std.ArrayList(ExpectedType).init(bun.default_allocator);
-        var dir = std.fs.Dir{ .fd = fd };
+        var dir = std.fs.Dir{ .fd = bun.fdcast(fd) };
         var iterator = DirIterator.iterate(dir);
         var entry = iterator.next();
         while (switch (entry) {
@@ -4734,6 +4788,9 @@ pub const NodeFS = struct {
         return Maybe(Return.Rm).todo;
     }
     pub fn stat(this: *NodeFS, args: Arguments.Stat, comptime flavor: Flavor) Maybe(Return.Stat) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Stat).todo;
+        }
         _ = flavor;
 
         return @as(Maybe(Return.Stat), switch (Syscall.stat(
@@ -4752,6 +4809,10 @@ pub const NodeFS = struct {
     }
 
     pub fn symlink(this: *NodeFS, args: Arguments.Symlink, comptime flavor: Flavor) Maybe(Return.Symlink) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Symlink).todo;
+        }
+
         var to_buf: [bun.MAX_PATH_BYTES]u8 = undefined;
 
         switch (comptime flavor) {
@@ -4767,6 +4828,10 @@ pub const NodeFS = struct {
         return Maybe(Return.Symlink).todo;
     }
     fn _truncate(this: *NodeFS, path: PathLike, len: JSC.WebCore.Blob.SizeType, comptime flavor: Flavor) Maybe(Return.Truncate) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Truncate).todo;
+        }
+
         switch (comptime flavor) {
             .sync => {
                 return Maybe(Return.Truncate).errno(C.truncate(path.sliceZ(&this.sync_error_buf), len)) orelse
@@ -4791,6 +4856,10 @@ pub const NodeFS = struct {
         };
     }
     pub fn unlink(this: *NodeFS, args: Arguments.Unlink, comptime flavor: Flavor) Maybe(Return.Unlink) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Unlink).todo;
+        }
+
         switch (comptime flavor) {
             .sync => {
                 return Maybe(Return.Unlink).errnoSysP(system.unlink(args.path.sliceZ(&this.sync_error_buf)), .unlink, args.path.slice()) orelse
@@ -4805,6 +4874,10 @@ pub const NodeFS = struct {
         return Maybe(Return.UnwatchFile).todo;
     }
     pub fn utimes(this: *NodeFS, args: Arguments.Utimes, comptime flavor: Flavor) Maybe(Return.Utimes) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Utimes).todo;
+        }
+
         var times = [2]std.c.timeval{
             .{
                 .tv_sec = args.mtime,
@@ -4833,6 +4906,10 @@ pub const NodeFS = struct {
     }
 
     pub fn lutimes(this: *NodeFS, args: Arguments.Lutimes, comptime flavor: Flavor) Maybe(Return.Lutimes) {
+        if (comptime Environment.isWindows) {
+            return Maybe(Return.Lutimes).todo;
+        }
+
         var times = [2]std.c.timeval{
             .{
                 .tv_sec = args.mtime,
